@Fryguy 2016-08-01
Can you expand on the OP description a bit...those number are meaningless to me without context (e.g. is that 142s total time or 142s faster but what percentage is that)

Do you have some benchmarks for "Speed[ing] up the time it takes to lookup a vm's parents"?