@Fryguy 2016-08-01
Overall code-wise looks good, but would like some benchmarks to justify it better.