@Fryguy 2016-08-01
Thanks for the numbers @kbrock .  Sounds good.  Only thing left is my comment on separate method and getting the tests to go green