@kbrock 2016-08-02
@Fryguy added scopes and cleared up dumb bug I introduced in fixing up cops