@Fryguy 2016-08-02
@kbrock I see you made a lot more changes after I gave a :+1:,  and started moving more stuff around...the code was good as is, but now I have to re-review in light of all the new changes...were they necessary to move for this PR?