@kbrock 2016-08-02
@Fryguy I'm sorry. The only changes that I think I made are related to your comments.

1. added a `filtered` scope per your suggestion
2. removed `filtered` helper method from specs per your suggestion
3. removed `assignment_mixin.rb` flat_map refactor that was unrelated, unnecessary, and was breaking the tests.