@kbrock 2016-08-05
testing:

I ran for >10,000 Vms, with the old method and the new method. and they returned the same value. (not conclusive, but a little reassuring)
Though the Vms probably had the same parent lineage anyway