@Fryguy 2016-08-05
@kbrock If you remove the parent_blue_folder commit, then this would be fine to merge...that particular commit has issues that could probably be done in a separate PR.

I assume the numbers in the OP are without that commit?  If so, they seemed good to me without that parent_blue_folder commit.