@kbrock 2016-08-06
status: extracged 10299, and probably discuss again the blue folder stuff with @Fryguy