@kbrock 2016-08-09
majority of the meat has been extracted.
What remains is moving `parent_blue_folder` over to using scopes. This reduces 8 (12 rows) queries to 3 (6 rows) and shows most of the benefit