@kbrock 2016-08-17
@Fryguy is there any legwork I can do on my side to make this easier to merge?
Or anything to extract to a separate PR?

thnx