@Fryguy 2016-09-09
@kbrock Would like to sit down with you to understand this one like we did last time.  I liked that last time we came to nice elegant answer that covered all use cases, and I have a feeling the same thing will happen here.