@kbrock 2016-11-30
WIP until I rebase and produce numbers again